clear cinn op lowering imp
