+++ paddle/cinn/hlir/framework/op_strategy.cc
@@ -13,9 +13,53 @@
 // limitations under the License.
 
 #include "paddle/cinn/hlir/framework/op_strategy.h"
+
+namespace {
+
+struct PyBindNodeAttrVisitor {
+  std::stringstream& out;
+  explicit PyBindNodeAttrVisitor(std::stringstream& out) : out(out) {}
+
+  void operator()(int v) { out << "int: " << v; }
+  void operator()(int64_t v) { out << "int64_t: " << v; }
+  void operator()(float v) { out << "float: " << v; }
+  void operator()(double v) { out << "double: " << v; }
+  void operator()(bool v) { out << "bool: " << v; }
+  void operator()(const std::string& v) { out << "string: " << v; }
+#define VISIT_ELEMENTS(T__)                                      \
+  void operator()(const std::vector<T__>& vs) {                  \
+    if (vs.empty()) return;                                      \
+    for (int i = 0; i < vs.size() - 1; i++) out << vs[i] << ","; \
+    out << vs.back();                                            \
+  }
+  VISIT_ELEMENTS(int)
+  VISIT_ELEMENTS(int64_t)
+  VISIT_ELEMENTS(float)
+  VISIT_ELEMENTS(double)
+  VISIT_ELEMENTS(bool)
+  VISIT_ELEMENTS(std::string)
+  VISIT_ELEMENTS(symbol::DimExpr)
+  VISIT_ELEMENTS(cinn::dialect::SymbolBinding)
+};
+
+}  // namespace
+
 namespace cinn {
 namespace hlir {
 namespace framework {
+
+std::ostream& operator<<(std::ostream& os, const NodeAttr& node_attr) {
+  std::stringstream ss;
+  ss << "NodeAttr:\n";
+  for (auto& item : node_attr.attr_store) {
+    std::stringstream os;
+    PyBindNodeAttrVisitor visitor(os);
+    absl::visit(visitor, item.second);
+    ss << "- " << os.str() << "\n";
+  }
+  os << ss.str();
+  return os;
+}
 
 std::shared_ptr<OpImpl> OpStrategy::SelectImpl(
     const std::shared_ptr<OpStrategy>& strategy) {
