+++ paddle/cinn/hlir/framework/op_strategy.h
@@ -18,9 +18,10 @@
 #include <utility>
 #include <vector>
 
-#include "paddle/cinn/hlir/framework/node.h"
+#include "paddle/cinn/hlir/framework/op.h"
 #include "paddle/cinn/hlir/framework/schedule.h"
 #include "paddle/cinn/lang/packed_func.h"
+#include "paddle/cinn/utils/type_defs.h"
 
 namespace cinn {
 namespace hlir {
@@ -30,6 +31,32 @@ using CINNCompute = lang::PackedFunc;
 using CINNSchedule = lang::PackedFunc;
 
 class OpStrategy;
+
+/**
+ * \brief Attributes of each node in graph.
+ *  The attributes include the node's name, the corresponding operator
+ *  and other parameters like axis.
+ */
+struct NodeAttr {
+  using attr_t = utils::Attribute;
+
+  /**
+   * \brief The operator this node uses.
+   */
+  const Operator* op{nullptr};
+
+  /**
+   * \brief The name of this node.
+   */
+  std::string node_name;
+
+  /**
+   * \brief The attributes stored as string in dictionary.
+   */
+  absl::flat_hash_map<std::string, attr_t> attr_store;
+};
+
+std::ostream& operator<<(std::ostream& os, const NodeAttr& node_attr);
 
 using StrategyFunction = std::function<std::shared_ptr<OpStrategy>(
     const NodeAttr&,
@@ -44,9 +71,6 @@ using StrategyFunctionSymbolic = std::function<std::shared_ptr<OpStrategy>(
     const std::vector<Type>&,
     const std::vector<std::vector<ir::Dim>>&,
     const cinn::common::Target&)>;
-
-using InferShapeFunction = std::function<std::vector<std::vector<int>>(
-    const std::vector<std::vector<int>>&, const AttrMapType&)>;
 
 //! Operator implementation that includes compute and schedule function.
 class OpImpl : public cinn::common::Object {
