+++ paddle/cinn/hlir/framework/pir/op_lowering_impl.cc
@@ -27,7 +27,6 @@
 #include "paddle/cinn/hlir/framework/pir/op_lowering_util.h"
 #include "paddle/cinn/hlir/framework/pir/trivial_op_impl.h"
 #include "paddle/cinn/hlir/framework/pir/utils.h"
-#include "paddle/cinn/hlir/op/external_api_registry.h"
 #include "paddle/cinn/hlir/pe/map_expr_to_ir.h"
 #include "paddle/cinn/ir/dim.h"
 #include "paddle/cinn/ir/group_schedule/base_group_scheduler.h"
@@ -61,7 +60,6 @@ namespace framework {
 namespace pir {
 
 using cinn::common::Type;
-using cinn::hlir::op::ExternalApiRegistry;
 using framework::OpPatternKind;
 using framework::StrategyFunction;
