+++ paddle/cinn/hlir/framework/pir/op_lowering_util.cc
@@ -142,199 +142,207 @@ using Visitor = std::function<std::vector<::pir::Operation*>(
   }
 }
 
-std::unordered_map<::pir::Operation*, ::pir::Operation*> BuildVirtualConsumer(
-    const GroupPtr& group) {
-  std::unordered_map<::pir::Operation*, ::pir::Operation*> virtual_consumers;
-  std::unordered_set<::pir::Operation*> ops_set(group->ops.begin(),
-                                                group->ops.end());
-  if (group->op_pattern_kind != framework::kReduction) {
-    return virtual_consumers;
-  }
+// std::unordered_map<::pir::Operation*, ::pir::Operation*>
+// BuildVirtualConsumer(
+//     const GroupPtr& group) {
+//   std::unordered_map<::pir::Operation*, ::pir::Operation*> virtual_consumers;
+//   std::unordered_set<::pir::Operation*> ops_set(group->ops.begin(),
+//                                                 group->ops.end());
+//   if (group->op_pattern_kind != framework::kReduction) {
+//     return virtual_consumers;
+//   }
 
-  ::pir::Operation* e_op = nullptr;
-  ::pir::Operation* r_op = nullptr;
-  for (auto master_op : group->master_ops) {
-    if (CompatibleInfo::OpKind(*master_op) != framework::kReduction) {
-      // producer exits reduce-sum and not consumers.
-      if (!e_op && FindReducerInRoute(master_op, ops_set, GetProducersInSet) &&
-          GetConsumersInSet(master_op, ops_set).size() == 0) {
-        e_op = master_op;
-      }
-    } else if (!r_op) {
-      r_op = master_op;
-    }
-  }
+//   ::pir::Operation* e_op = nullptr;
+//   ::pir::Operation* r_op = nullptr;
+//   for (auto master_op : group->master_ops) {
+//     if (CompatibleInfo::OpKind(*master_op) != framework::kReduction) {
+//       // producer exits reduce-sum and not consumers.
+//       if (!e_op && FindReducerInRoute(master_op, ops_set, GetProducersInSet)
+//       &&
+//           GetConsumersInSet(master_op, ops_set).size() == 0) {
+//         e_op = master_op;
+//       }
+//     } else if (!r_op) {
+//       r_op = master_op;
+//     }
+//   }
 
-  // try to find reducer with different shape.
-  for (auto output_op : group->output_ops) {
-    if (CompatibleInfo::OpKind(*output_op) == framework::kReduction) {
-      if (isl_ast_expr_op_and_then) {
-        virtual_consumers[output_op] = e_op;
-      }
-      continue;
-    }
-    if (FindNearestReducer(output_op, ops_set)) {
-      continue;
-    }
+//   // try to find reducer with different shape.
+//   for (auto output_op : group->output_ops) {
+//     if (CompatibleInfo::OpKind(*output_op) == framework::kReduction) {
+//       if (isl_ast_expr_op_and_then) {
+//         virtual_consumers[output_op] = e_op;
+//       }
+//       continue;
+//     }
+//     if (FindNearestReducer(output_op, ops_set)) {
+//       continue;
+//     }
 
-    bool found = false;
-    std::unordered_set<::pir::Operation*> visited;
-    std::queue<::pir::Operation*> candidates;
+//     bool found = false;
+//     std::unordered_set<::pir::Operation*> visited;
+//     std::queue<::pir::Operation*> candidates;
 
-    candidates.push(output_op);
-    visited.insert(output_op);
-    // from producers find reducer consumer.
-    while (!found && !candidates.empty()) {
-      auto candidate = candidates.front();
-      candidates.pop();
+//     candidates.push(output_op);
+//     visited.insert(output_op);
+//     // from producers find reducer consumer.
+//     while (!found && !candidates.empty()) {
+//       auto candidate = candidates.front();
+//       candidates.pop();
 
-      for (auto producer : GetProducersInSet(candidate, ops_set)) {
-        if (visited.count(producer)) {
-          continue;
-        }
+//       for (auto producer : GetProducersInSet(candidate, ops_set)) {
+//         if (visited.count(producer)) {
+//           continue;
+//         }
 
-        auto reducer = FindReducerInRoute(producer, ops_set, GetConsumersInSet);
-        if (reducer) {
-          virtual_consumers[output_op] = reducer;
-          found = true;
-          break;
-        }
-        candidates.push(producer);
-        visited.insert(producer);
-      }
-    }
+//         auto reducer = FindReducerInRoute(producer, ops_set,
+//         GetConsumersInSet); if (reducer) {
+//           virtual_consumers[output_op] = reducer;
+//           found = true;
+//           break;
+//         }
+//         candidates.push(producer);
+//         visited.insert(producer);
+//       }
+//     }
 
-    auto output_shape = CompatibleInfo::ValueShape(output_op->result(0));
-    if (!found && output_op != e_op && e_op) {
-      auto e_output_shape = CompatibleInfo::ValueShape(e_op->result(0));
-      if (CompatibleInfo::ShapeProduct(output_shape) ==
-          CompatibleInfo::ShapeProduct(e_output_shape)) {
-        virtual_consumers[output_op] = e_op;
-        found = true;
-      }
-    }
-    if (!found && r_op) {
-      auto r_input_shape = CompatibleInfo::ValueShape(r_op->operand_source(0));
-      if (CompatibleInfo::ShapeProduct(output_shape) ==
-          CompatibleInfo::ShapeProduct(r_input_shape)) {
-        virtual_consumers[output_op] = r_op;
-        found = true;
-      }
-    }
-  }
-  // Establish virtual consumer relationships between output nodes with the same
-  // shape. This allows the calculation of output nodes without affiliation to
-  // be placed under the same loop.
-  std::unordered_map<int, ::pir::Operation*> numel_consumers;
-  for (auto out_op : group->output_ops) {
-    if (virtual_consumers.find(out_op) != virtual_consumers.end() ||
-        !GetConsumersInSet(out_op, ops_set).empty()) {
-      continue;
-    }
-    auto shape = CompatibleInfo::ValueShape(out_op->result(0));
-    int numel = CompatibleInfo::ShapeProduct(shape);
-    if (numel_consumers.find(numel) == numel_consumers.end()) {
-      numel_consumers.insert(std::make_pair(numel, out_op));
-    } else {
-      virtual_consumers[out_op] = numel_consumers[numel];
-    }
-  }
-  return virtual_consumers;
-}
+//     auto output_shape = CompatibleInfo::ValueShape(output_op->result(0));
+//     if (!found && output_op != e_op && e_op) {
+//       auto e_output_shape = CompatibleInfo::ValueShape(e_op->result(0));
+//       if (CompatibleInfo::ShapeProduct(output_shape) ==
+//           CompatibleInfo::ShapeProduct(e_output_shape)) {
+//         virtual_consumers[output_op] = e_op;
+//         found = true;
+//       }
+//     }
+//     if (!found && r_op) {
+//       auto r_input_shape =
+//       CompatibleInfo::ValueShape(r_op->operand_source(0)); if
+//       (CompatibleInfo::ShapeProduct(output_shape) ==
+//           CompatibleInfo::ShapeProduct(r_input_shape)) {
+//         virtual_consumers[output_op] = r_op;
+//         found = true;
+//       }
+//     }
+//   }
+//   // Establish virtual consumer relationships between output nodes with the
+//   same
+//   // shape. This allows the calculation of output nodes without affiliation
+//   to
+//   // be placed under the same loop.
+//   std::unordered_map<int, ::pir::Operation*> numel_consumers;
+//   for (auto out_op : group->output_ops) {
+//     if (virtual_consumers.find(out_op) != virtual_consumers.end() ||
+//         !GetConsumersInSet(out_op, ops_set).empty()) {
+//       continue;
+//     }
+//     auto shape = CompatibleInfo::ValueShape(out_op->result(0));
+//     int numel = CompatibleInfo::ShapeProduct(shape);
+//     if (numel_consumers.find(numel) == numel_consumers.end()) {
+//       numel_consumers.insert(std::make_pair(numel, out_op));
+//     } else {
+//       virtual_consumers[out_op] = numel_consumers[numel];
+//     }
+//   }
+//   return virtual_consumers;
+// }
 
-std::vector<::pir::Operation*> BFSTopologicalOrderWithPriority(
-    const GroupPtr& group,
-    const std::unordered_map<::pir::Operation*, ::pir::Operation*>&
-        virtual_consumers) {
-  struct OpWithPriority {
-    ::pir::Operation* op;
-    int priority;
-  };
+// std::vector<::pir::Operation*> BFSTopologicalOrderWithPriority(
+//     const GroupPtr& group,
+//     const std::unordered_map<::pir::Operation*, ::pir::Operation*>&
+//         virtual_consumers) {
+//   struct OpWithPriority {
+//     ::pir::Operation* op;
+//     int priority;
+//   };
 
-  struct Comparator {
-    bool operator()(const OpWithPriority& lhs, const OpWithPriority& rhs) {
-      return lhs.priority > rhs.priority;
-    }
-  };
+//   struct Comparator {
+//     bool operator()(const OpWithPriority& lhs, const OpWithPriority& rhs) {
+//       return lhs.priority > rhs.priority;
+//     }
+//   };
 
-  std::vector<::pir::Operation*> ops_in_order;
-  std::unordered_set<::pir::Operation*> visited;
-  std::unordered_set<::pir::Operation*> ops_set(group->ops.begin(),
-                                                group->ops.end());
-  std::unordered_map<::pir::Operation*, int> degree_map;
-  std::priority_queue<OpWithPriority, std::vector<OpWithPriority>, Comparator>
-      priority_candidates;
-  std::vector<int> visited_numel;
+//   std::vector<::pir::Operation*> ops_in_order;
+//   std::unordered_set<::pir::Operation*> visited;
+//   std::unordered_set<::pir::Operation*> ops_set(group->ops.begin(),
+//                                                 group->ops.end());
+//   std::unordered_map<::pir::Operation*, int> degree_map;
+//   std::priority_queue<OpWithPriority, std::vector<OpWithPriority>,
+//   Comparator>
+//       priority_candidates;
+//   std::vector<int> visited_numel;
 
-  // Calculate the priority of a node.
-  // The smaller the value, the higher the priority.
-  // Prioritize the same shape before considering OpPattern
-  auto PriorityFunc = [&visited_numel](::pir::Operation* op) -> int {
-    auto op_shape = CompatibleInfo::ValueShape(op->result(0));
-    int numel = CompatibleInfo::ShapeProduct(op_shape);
-    int index = -1;
-    for (int i = 0; i < visited_numel.size(); ++i) {
-      if (numel == visited_numel[i]) {
-        index = i;
-        break;
-      }
-    }
-    if (index == -1) {
-      index = visited_numel.size();
-      visited_numel.push_back(numel);
-    }
-    return index * 10 + static_cast<int>(CompatibleInfo::OpKind(*op));
-  };
+//   // Calculate the priority of a node.
+//   // The smaller the value, the higher the priority.
+//   // Prioritize the same shape before considering OpPattern
+//   auto PriorityFunc = [&visited_numel](::pir::Operation* op) -> int {
+//     auto op_shape = CompatibleInfo::ValueShape(op->result(0));
+//     int numel = CompatibleInfo::ShapeProduct(op_shape);
+//     int index = -1;
+//     for (int i = 0; i < visited_numel.size(); ++i) {
+//       if (numel == visited_numel[i]) {
+//         index = i;
+//         break;
+//       }
+//     }
+//     if (index == -1) {
+//       index = visited_numel.size();
+//       visited_numel.push_back(numel);
+//     }
+//     return index * 10 + static_cast<int>(CompatibleInfo::OpKind(*op));
+//   };
 
-  for (::pir::Operation* op : ops_set) {
-    auto consumers = FindConsumers(op, ops_set, virtual_consumers);
-    // Some nodes may have multiple edges between them, resulting in duplicates
-    // in the consumer. We only need to calculate once.
-    std::unordered_set<::pir::Operation*> consumers_without_duplicate(
-        consumers.begin(), consumers.end());
-    degree_map[op] = consumers_without_duplicate.size();
-    if (degree_map.at(op) == 0) {
-      priority_candidates.push(OpWithPriority{op, PriorityFunc(op)});
-    }
-  }
+//   for (::pir::Operation* op : ops_set) {
+//     auto consumers = FindConsumers(op, ops_set, virtual_consumers);
+//     // Some nodes may have multiple edges between them, resulting in
+//     duplicates
+//     // in the consumer. We only need to calculate once.
+//     std::unordered_set<::pir::Operation*> consumers_without_duplicate(
+//         consumers.begin(), consumers.end());
+//     degree_map[op] = consumers_without_duplicate.size();
+//     if (degree_map.at(op) == 0) {
+//       priority_candidates.push(OpWithPriority{op, PriorityFunc(op)});
+//     }
+//   }
 
-  // Nested BFS, outer layer traverses priority, inner layer performs BFS on
-  // current priority.
-  while (!priority_candidates.empty()) {
-    ::pir::Operation* cur_priority_op = priority_candidates.top().op;
-    priority_candidates.pop();
+//   // Nested BFS, outer layer traverses priority, inner layer performs BFS on
+//   // current priority.
+//   while (!priority_candidates.empty()) {
+//     ::pir::Operation* cur_priority_op = priority_candidates.top().op;
+//     priority_candidates.pop();
 
-    std::queue<::pir::Operation*> bfs_queue;
-    bfs_queue.push(cur_priority_op);
-    visited.insert(cur_priority_op);
-    while (!bfs_queue.empty()) {
-      ::pir::Operation* cur = bfs_queue.front();
-      bfs_queue.pop();
+//     std::queue<::pir::Operation*> bfs_queue;
+//     bfs_queue.push(cur_priority_op);
+//     visited.insert(cur_priority_op);
+//     while (!bfs_queue.empty()) {
+//       ::pir::Operation* cur = bfs_queue.front();
+//       bfs_queue.pop();
 
-      ops_in_order.push_back(cur);
-      auto producers = FindProducers(cur, ops_set, virtual_consumers);
-      std::unordered_set<::pir::Operation*> producers_without_duplicate(
-          producers.begin(), producers.end());
-      for (::pir::Operation* op : producers_without_duplicate) {
-        --degree_map[op];
-        // Ensure that each node is accessed only once and maintain topological
-        // order.
-        if (visited.count(op) != 0 || degree_map[op] != 0) {
-          continue;
-        }
-        // Perform BFS access to the current priority producers
-        int op_priority = PriorityFunc(op);
-        if (op_priority <= PriorityFunc(cur_priority_op)) {
-          bfs_queue.push(op);
-          visited.insert(op);
-        } else {
-          priority_candidates.push(OpWithPriority{op, op_priority});
-        }
-      }
-    }
-  }
-  return ops_in_order;
-}
+//       ops_in_order.push_back(cur);
+//       auto producers = FindProducers(cur, ops_set, virtual_consumers);
+//       std::unordered_set<::pir::Operation*> producers_without_duplicate(
+//           producers.begin(), producers.end());
+//       for (::pir::Operation* op : producers_without_duplicate) {
+//         --degree_map[op];
+//         // Ensure that each node is accessed only once and maintain
+//         topological
+//         // order.
+//         if (visited.count(op) != 0 || degree_map[op] != 0) {
+//           continue;
+//         }
+//         // Perform BFS access to the current priority producers
+//         int op_priority = PriorityFunc(op);
+//         if (op_priority <= PriorityFunc(cur_priority_op)) {
+//           bfs_queue.push(op);
+//           visited.insert(op);
+//         } else {
+//           priority_candidates.push(OpWithPriority{op, op_priority});
+//         }
+//       }
+//     }
+//   }
+//   return ops_in_order;
+// }
 
 std::unordered_set<::pir::Operation*> GetMasters(
     ::pir::Operation* op,
@@ -374,59 +382,59 @@ bool IsConstOp(const ::pir::Operation* op) {
   return const_op_type.count(CompatibleInfo::OpName(*op));
 }
 
-bool CanbeInline(::pir::Operation* op,
-                 ::pir::Operation* reducer,
-                 PrettyNamer* pretty_name,
-                 const std::vector<::pir::Operation*> consumers,
-                 const std::unordered_set<::pir::Operation*> masters,
-                 const GroupPtr& group,
-                 const std::unordered_set<::pir::Operation*>& ops_set) {
-  if (group->output_ops.count(op)) {
-    return false;
-  }
-  for (auto consumer : consumers) {
-    if (CompatibleInfo::OpKind(*consumer) == framework::kReduction) {
-      return false;
-    }
-  }
+// bool CanbeInline(::pir::Operation* op,
+//                  ::pir::Operation* reducer,
+//                  PrettyNamer* pretty_name,
+//                  const std::vector<::pir::Operation*> consumers,
+//                  const std::unordered_set<::pir::Operation*> masters,
+//                  const GroupPtr& group,
+//                  const std::unordered_set<::pir::Operation*>& ops_set) {
+//   if (group->output_ops.count(op)) {
+//     return false;
+//   }
+//   for (auto consumer : consumers) {
+//     if (CompatibleInfo::OpKind(*consumer) == framework::kReduction) {
+//       return false;
+//     }
+//   }
 
-  if (IsConstOp(op)) {
-    return true;
-  }
-  if (CompatibleInfo::OpKind(*op) == framework::kReduction) {
-    return false;
-  }
+//   if (IsConstOp(op)) {
+//     return true;
+//   }
+//   if (CompatibleInfo::OpKind(*op) == framework::kReduction) {
+//     return false;
+//   }
 
-  if (consumers.size() == 1) {
-    return true;
-  }
-  auto op_shape = CompatibleInfo::ValueShape(op->result(0));
-  if (reducer) {
-    // node is before reducer and node is not after reduce.
-    if (FindReducerInRoute(op, ops_set, GetConsumersInSet) &&
-        !FindReducerInRoute(op, ops_set, GetProducersInSet)) {
-      auto input_shape = CompatibleInfo::ValueShape(reducer->result(0));
-      // check with same shape with reducer input.
-      if (CompatibleInfo::ShapeProduct(op_shape) !=
-          CompatibleInfo::ShapeProduct(input_shape)) {
-        return true;
-      }
-    }
+//   if (consumers.size() == 1) {
+//     return true;
+//   }
+//   auto op_shape = CompatibleInfo::ValueShape(op->result(0));
+//   if (reducer) {
+//     // node is before reducer and node is not after reduce.
+//     if (FindReducerInRoute(op, ops_set, GetConsumersInSet) &&
+//         !FindReducerInRoute(op, ops_set, GetProducersInSet)) {
+//       auto input_shape = CompatibleInfo::ValueShape(reducer->result(0));
+//       // check with same shape with reducer input.
+//       if (CompatibleInfo::ShapeProduct(op_shape) !=
+//           CompatibleInfo::ShapeProduct(input_shape)) {
+//         return true;
+//       }
+//     }
 
-    return false;
-  } else {
-    auto op_shape_size = CompatibleInfo::ShapeProduct(op_shape);
-    for (auto master : masters) {
-      auto master_shape = CompatibleInfo::ValueShape(master->result(0));
-      auto master_size = CompatibleInfo::ShapeProduct(master_shape);
-      if (op_shape_size != master_size) {
-        return true;
-      }
-    }
+//     return false;
+//   } else {
+//     auto op_shape_size = CompatibleInfo::ShapeProduct(op_shape);
+//     for (auto master : masters) {
+//       auto master_shape = CompatibleInfo::ValueShape(master->result(0));
+//       auto master_size = CompatibleInfo::ShapeProduct(master_shape);
+//       if (op_shape_size != master_size) {
+//         return true;
+//       }
+//     }
 
-    return false;
-  }
-}
+//     return false;
+//   }
+// }
 
 ::pir::Operation* GetMasterToComputeAt(
     ::pir::Operation* op,
@@ -1260,67 +1268,68 @@ void MergeReduceLoop(
   } while (--index >= 0);
 }
 
-void LoopComputeAt(
-    ir::IRSchedule& ir_sch,  // NOLINT
-    ::pir::Operation* op,
-    ::pir::Operation* master,
-    PrettyNamer* pretty_name,
-    const GroupPtr& group,
-    const std::unordered_map<::pir::Value, ir::Tensor>& tensor_map,
-    const std::unordered_map<std::string, ir::Tensor>& tmp_tensor_info) {
-  auto op_out_name =
-      pretty_name->GetOrNew(op->result(0), CompatibleInfo::kNamePrefix);
-  if (!group->output_ops.count(op)) {
-    auto block = ir_sch.GetBlock(op_out_name);
-    ir_sch.SetBuffer(block, "local");
-  }
+// void LoopComputeAt(
+//     ir::IRSchedule& ir_sch,  // NOLINT
+//     ::pir::Operation* op,
+//     ::pir::Operation* master,
+//     PrettyNamer* pretty_name,
+//     const GroupPtr& group,
+//     const std::unordered_map<::pir::Value, ir::Tensor>& tensor_map,
+//     const std::unordered_map<std::string, ir::Tensor>& tmp_tensor_info) {
+//   auto op_out_name =
+//       pretty_name->GetOrNew(op->result(0), CompatibleInfo::kNamePrefix);
+//   if (!group->output_ops.count(op)) {
+//     auto block = ir_sch.GetBlock(op_out_name);
+//     ir_sch.SetBuffer(block, "local");
+//   }
 
-  if (CompatibleInfo::OpKind(*op) == framework::kReduction) {
-    MergeReduceLoop(
-        ir_sch, op, master, pretty_name, tensor_map, tmp_tensor_info);
-    return;
-  }
+//   if (CompatibleInfo::OpKind(*op) == framework::kReduction) {
+//     MergeReduceLoop(
+//         ir_sch, op, master, pretty_name, tensor_map, tmp_tensor_info);
+//     return;
+//   }
 
-  if (op == master) return;
-  auto master_out = master->result(0);
-  auto master_out_name =
-      pretty_name->GetOrNew(master_out, CompatibleInfo::kNamePrefix);
+//   if (op == master) return;
+//   auto master_out = master->result(0);
+//   auto master_out_name =
+//       pretty_name->GetOrNew(master_out, CompatibleInfo::kNamePrefix);
 
-  auto node_loops = ir_sch.GetLoops(op_out_name);
-  auto master_loops = ir_sch.GetLoops(master_out_name);
+//   auto node_loops = ir_sch.GetLoops(op_out_name);
+//   auto master_loops = ir_sch.GetLoops(master_out_name);
 
-  if (CompatibleInfo::OpKind(*master) == framework::kReduction) {
-    // find real master loops.
-    std::string prefix = "", post = "";
-    for (int idx = 0;; ++idx) {
-      if (!tmp_tensor_info.count(master_out_name + post)) {
-        break;
-      }
-      auto tensor = tmp_tensor_info.at(master_out_name + post);
-      if (!ir_sch.HasBlock(tensor->name)) {
-        break;
-      }
+//   if (CompatibleInfo::OpKind(*master) == framework::kReduction) {
+//     // find real master loops.
+//     std::string prefix = "", post = "";
+//     for (int idx = 0;; ++idx) {
+//       if (!tmp_tensor_info.count(master_out_name + post)) {
+//         break;
+//       }
+//       auto tensor = tmp_tensor_info.at(master_out_name + post);
+//       if (!ir_sch.HasBlock(tensor->name)) {
+//         break;
+//       }
 
-      prefix = post;
-      post = "_" + std::to_string(idx);
-    }
-    auto tensor = tmp_tensor_info.at(master_out_name + prefix);
-    master_loops = ir_sch.GetLoops(tensor->name);
-  }
+//       prefix = post;
+//       post = "_" + std::to_string(idx);
+//     }
+//     auto tensor = tmp_tensor_info.at(master_out_name + prefix);
+//     master_loops = ir_sch.GetLoops(tensor->name);
+//   }
 
-  int index = std::min(node_loops.size(), master_loops.size()) - 1;
-  do {
-    // if loop range is not equal.
-    if (node_loops[index].As<ir::For>()->extent.as_int32() !=
-        master_loops[index].As<ir::For>()->extent.as_int32()) {
-      continue;
-    }
-    MergeLoops(
-        ir_sch.GetModule().GetExprs().at(0), node_loops, master_loops, index);
+//   int index = std::min(node_loops.size(), master_loops.size()) - 1;
+//   do {
+//     // if loop range is not equal.
+//     if (node_loops[index].As<ir::For>()->extent.as_int32() !=
+//         master_loops[index].As<ir::For>()->extent.as_int32()) {
+//       continue;
+//     }
+//     MergeLoops(
+//         ir_sch.GetModule().GetExprs().at(0), node_loops, master_loops,
+//         index);
 
-    break;
-  } while (--index >= 0);
-}
+//     break;
+//   } while (--index >= 0);
+// }
 
 void LoopAssignReduce(
     ir::IRSchedule& ir_sch,  // NOLINT
@@ -1547,86 +1556,86 @@ std::unordered_map<std::string, ::pir::Value> GetOutValueSet(
   return out_value_set;
 }
 
-void SyncThreadWithShared(
-    ir::IRSchedule& ir_sch,  // NOLINT
-    const GroupPtr& group,
-    PrettyNamer* pretty_name,
-    const std::unordered_set<::pir::Operation*>& ops_inline,
-    const std::unordered_set<::pir::Operation*>& ops_set,
-    const std::unordered_map<::pir::Value, ir::Tensor>& tensor_map) {
-  auto exprs_inorder = ir_sch.GetAllBlocks();
-  auto op_out_set = GetOutValueSet(pretty_name, ops_set);
+// void SyncThreadWithShared(
+//     ir::IRSchedule& ir_sch,  // NOLINT
+//     const GroupPtr& group,
+//     PrettyNamer* pretty_name,
+//     const std::unordered_set<::pir::Operation*>& ops_inline,
+//     const std::unordered_set<::pir::Operation*>& ops_set,
+//     const std::unordered_map<::pir::Value, ir::Tensor>& tensor_map) {
+//   auto exprs_inorder = ir_sch.GetAllBlocks();
+//   auto op_out_set = GetOutValueSet(pretty_name, ops_set);
 
-  std::unordered_set<std::string> sync_mark;
-  auto check_sync_mark = [&](const int start, const std::string& m_id) {
-    for (int idx = start + 1; exprs_inorder.size(); ++idx) {
-      auto expr = exprs_inorder[idx];
-      CHECK(expr.As<ir::ScheduleBlockRealize>());
-      CHECK(expr.As<ir::ScheduleBlockRealize>()
-                ->schedule_block.As<ir::ScheduleBlock>());
-      auto block = expr.As<ir::ScheduleBlockRealize>()
-                       ->schedule_block.As<ir::ScheduleBlock>();
+//   std::unordered_set<std::string> sync_mark;
+//   auto check_sync_mark = [&](const int start, const std::string& m_id) {
+//     for (int idx = start + 1; exprs_inorder.size(); ++idx) {
+//       auto expr = exprs_inorder[idx];
+//       CHECK(expr.As<ir::ScheduleBlockRealize>());
+//       CHECK(expr.As<ir::ScheduleBlockRealize>()
+//                 ->schedule_block.As<ir::ScheduleBlock>());
+//       auto block = expr.As<ir::ScheduleBlockRealize>()
+//                        ->schedule_block.As<ir::ScheduleBlock>();
 
-      if (sync_mark.count(block->name)) {
-        return false;
-      }
+//       if (sync_mark.count(block->name)) {
+//         return false;
+//       }
 
-      if (block->name == m_id) {
-        return true;
-      }
-    }
-    return false;
-  };
+//       if (block->name == m_id) {
+//         return true;
+//       }
+//     }
+//     return false;
+//   };
 
-  for (int idx = 0; idx < exprs_inorder.size() - 1; ++idx) {
-    auto expr = exprs_inorder[idx];
-    CHECK(expr.As<ir::ScheduleBlockRealize>());
-    CHECK(expr.As<ir::ScheduleBlockRealize>()
-              ->schedule_block.As<ir::ScheduleBlock>());
-    auto block = expr.As<ir::ScheduleBlockRealize>()
-                     ->schedule_block.As<ir::ScheduleBlock>();
+//   for (int idx = 0; idx < exprs_inorder.size() - 1; ++idx) {
+//     auto expr = exprs_inorder[idx];
+//     CHECK(expr.As<ir::ScheduleBlockRealize>());
+//     CHECK(expr.As<ir::ScheduleBlockRealize>()
+//               ->schedule_block.As<ir::ScheduleBlock>());
+//     auto block = expr.As<ir::ScheduleBlockRealize>()
+//                      ->schedule_block.As<ir::ScheduleBlock>();
 
-    if (!op_out_set.count(block->name)) {
-      continue;
-    }
-    auto op_data = op_out_set.find(block->name)->second;
-    auto* op = op_data.defining_op();
-    auto op_shape = CompatibleInfo::ValueShape(op_data);
+//     if (!op_out_set.count(block->name)) {
+//       continue;
+//     }
+//     auto op_data = op_out_set.find(block->name)->second;
+//     auto* op = op_data.defining_op();
+//     auto op_shape = CompatibleInfo::ValueShape(op_data);
 
-    auto masters = GetMasters(op, pretty_name, ops_inline, ops_set);
-    if (masters.empty()) {
-      continue;
-    }
+//     auto masters = GetMasters(op, pretty_name, ops_inline, ops_set);
+//     if (masters.empty()) {
+//       continue;
+//     }
 
-    bool do_set_buffer_to_shared = false;
-    for (auto master : masters) {
-      auto master_data = master->result(0);
-      auto master_shape = CompatibleInfo::ValueShape(master_data);
-      if (CompatibleInfo::OpKind(*master) == framework::kReduction) {
-        master_shape = CompatibleInfo::ValueShape(master->operand_source(0));
-      }
+//     bool do_set_buffer_to_shared = false;
+//     for (auto master : masters) {
+//       auto master_data = master->result(0);
+//       auto master_shape = CompatibleInfo::ValueShape(master_data);
+//       if (CompatibleInfo::OpKind(*master) == framework::kReduction) {
+//         master_shape = CompatibleInfo::ValueShape(master->operand_source(0));
+//       }
 
-      auto op_shape_size = CompatibleInfo::ShapeProduct(op_shape);
-      auto master_shape_size = CompatibleInfo::ShapeProduct(master_shape);
-      std::string master_data_name =
-          pretty_name->GetOrNew(master_data, CompatibleInfo::kNamePrefix);
-      if (op_shape_size != master_shape_size) {
-        if (check_sync_mark(idx, master_data_name)) {
-          auto loops = ir_sch.GetLoops(master_data_name);
-          ir_sch.SyncThreads(loops.back(), false);
-          sync_mark.insert(master_data_name);
-        }
-        do_set_buffer_to_shared = true;
-      }
-    }
-    if (do_set_buffer_to_shared &&
-        group->output_ops.find(op) == group->output_ops.end()) {
-      auto block = ir_sch.GetBlock(
-          pretty_name->GetOrNew(op_data, CompatibleInfo::kNamePrefix));
-      ir_sch.SetBuffer(block, "shared");
-    }
-  }
-}
+//       auto op_shape_size = CompatibleInfo::ShapeProduct(op_shape);
+//       auto master_shape_size = CompatibleInfo::ShapeProduct(master_shape);
+//       std::string master_data_name =
+//           pretty_name->GetOrNew(master_data, CompatibleInfo::kNamePrefix);
+//       if (op_shape_size != master_shape_size) {
+//         if (check_sync_mark(idx, master_data_name)) {
+//           auto loops = ir_sch.GetLoops(master_data_name);
+//           ir_sch.SyncThreads(loops.back(), false);
+//           sync_mark.insert(master_data_name);
+//         }
+//         do_set_buffer_to_shared = true;
+//       }
+//     }
+//     if (do_set_buffer_to_shared &&
+//         group->output_ops.find(op) == group->output_ops.end()) {
+//       auto block = ir_sch.GetBlock(
+//           pretty_name->GetOrNew(op_data, CompatibleInfo::kNamePrefix));
+//       ir_sch.SetBuffer(block, "shared");
+//     }
+//   }
+// }
 
 }  // namespace pir
 }  // namespace framework
