+++ paddle/cinn/hlir/framework/pir/op_lowering_util.h
@@ -17,7 +17,6 @@
 #include <memory>
 #include <queue>
 
-#include "paddle/cinn/hlir/framework/pir/group.h"
 #include "paddle/cinn/hlir/framework/pir/op_lowering_group.h"
 #include "paddle/cinn/ir/schedule/ir_schedule.h"
 #include "paddle/cinn/ir/tensor.h"
@@ -26,13 +25,13 @@ namespace cinn {
 namespace hlir {
 namespace framework {
 namespace pir {
-using GroupPtr = std::shared_ptr<Group>;
 using OpLoweringGroupPtr = std::shared_ptr<OpLoweringGroup>;
 
 class PrettyNamer;
 
-std::unordered_map<::pir::Operation*, ::pir::Operation*> BuildVirtualConsumer(
-    const GroupPtr& group);
+// std::unordered_map<::pir::Operation*, ::pir::Operation*>
+// BuildVirtualConsumer(
+//     const GroupPtr& group);
 
 std::vector<::pir::Value*> GetAllNodeData(::pir::Operation* op);
 
@@ -43,15 +42,15 @@ bool IsConstOp(const ::pir::Operation* op);
 std::vector<::pir::Operation*> GetConsumersInSet(
     ::pir::Operation* op, const std::unordered_set<::pir::Operation*>& ops);
 
-std::vector<::pir::Operation*> TopologicalOrder(
-    const GroupPtr& group,
-    const std::unordered_map<::pir::Operation*, ::pir::Operation*>&
-        virtual_consumers);
+// std::vector<::pir::Operation*> TopologicalOrder(
+//     const GroupPtr& group,
+//     const std::unordered_map<::pir::Operation*, ::pir::Operation*>&
+//         virtual_consumers);
 
-std::vector<::pir::Operation*> BFSTopologicalOrderWithPriority(
-    const GroupPtr& group,
-    const std::unordered_map<::pir::Operation*, ::pir::Operation*>&
-        virtual_consumers);
+// std::vector<::pir::Operation*> BFSTopologicalOrderWithPriority(
+//     const GroupPtr& group,
+//     const std::unordered_map<::pir::Operation*, ::pir::Operation*>&
+//         virtual_consumers);
 
 ::pir::Operation* FindGlobalReducer(
     const std::vector<::pir::Operation*>& ops_in_order);
@@ -59,13 +58,13 @@ std::vector<::pir::Operation*> BFSTopologicalOrderWithPriority(
 ::pir::Operation* FindNearestReducer(
     ::pir::Operation* op, const std::unordered_set<::pir::Operation*>& ops_set);
 
-bool CanbeInline(::pir::Operation* op,
-                 ::pir::Operation* reducer,
-                 PrettyNamer* pretty_name,
-                 const std::vector<::pir::Operation*> consumers,
-                 const std::unordered_set<::pir::Operation*> masters,
-                 const GroupPtr& group,
-                 const std::unordered_set<::pir::Operation*>& ops_set);
+// bool CanbeInline(::pir::Operation* op,
+//                  ::pir::Operation* reducer,
+//                  PrettyNamer* pretty_name,
+//                  const std::vector<::pir::Operation*> consumers,
+//                  const std::unordered_set<::pir::Operation*> masters,
+//                  const GroupPtr& group,
+//                  const std::unordered_set<::pir::Operation*>& ops_set);
 
 ::pir::Operation* GetMasterToComputeAt(
     ::pir::Operation* op,
@@ -91,22 +90,22 @@ void LoopAssignReduce(
     const std::unordered_map<::pir::Value, ir::Tensor>& tensor_map,
     const std::unordered_map<std::string, ir::Tensor>& tmp_tensor_info);
 
-void LoopComputeAt(
-    ir::IRSchedule& ir_sch,  // NOLINT
-    ::pir::Operation* op,
-    ::pir::Operation* master,
-    PrettyNamer* pretty_name,
-    const GroupPtr& group,
-    const std::unordered_map<::pir::Value, ir::Tensor>& tensor_map,
-    const std::unordered_map<std::string, ir::Tensor>& tmp_tensor_info);
+// void LoopComputeAt(
+//     ir::IRSchedule& ir_sch,  // NOLINT
+//     ::pir::Operation* op,
+//     ::pir::Operation* master,
+//     PrettyNamer* pretty_name,
+//     const GroupPtr& group,
+//     const std::unordered_map<::pir::Value, ir::Tensor>& tensor_map,
+//     const std::unordered_map<std::string, ir::Tensor>& tmp_tensor_info);
 
-void SyncThreadWithShared(
-    ir::IRSchedule& ir_sch,  // NOLINT
-    const GroupPtr& group,
-    PrettyNamer* pretty_name,
-    const std::unordered_set<::pir::Operation*>& ops_inline,
-    const std::unordered_set<::pir::Operation*>& ops_set,
-    const std::unordered_map<::pir::Value, ir::Tensor>& tensor_map);
+// void SyncThreadWithShared(
+//     ir::IRSchedule& ir_sch,  // NOLINT
+//     const GroupPtr& group,
+//     PrettyNamer* pretty_name,
+//     const std::unordered_set<::pir::Operation*>& ops_inline,
+//     const std::unordered_set<::pir::Operation*>& ops_set,
+//     const std::unordered_map<::pir::Value, ir::Tensor>& tensor_map);
 
 }  // namespace pir
 }  // namespace framework
