+++ paddle/cinn/hlir/op/CMakeLists.txt
@@ -11,8 +11,4 @@ gather_srcs(
   elementwise.cc
   reduction.cc
   op_util.cc
-  custom_call.cc
-  external_api_registry.cc)
-
-cinn_cc_test(test_external_api_registry SRCS external_api_registry_test.cc DEPS
-             cinncore)
+  custom_call.cc)
