+++ paddle/cinn/hlir/op/contrib/gather_nd.cc
@@ -24,7 +24,6 @@
 #include "paddle/cinn/common/context.h"
 #include "paddle/cinn/common/macros.h"
 #include "paddle/cinn/common/type.h"
-#include "paddle/cinn/hlir/framework/node.h"
 #include "paddle/cinn/hlir/framework/op.h"
 #include "paddle/cinn/hlir/framework/op_strategy.h"
 #include "paddle/cinn/hlir/pe/elementwise.h"
