+++ paddle/cinn/hlir/op/contrib/reciprocal.cc
@@ -22,7 +22,6 @@
 #include "paddle/cinn/common/context.h"
 #include "paddle/cinn/common/macros.h"
 #include "paddle/cinn/common/target.h"
-#include "paddle/cinn/hlir/framework/node.h"
 #include "paddle/cinn/hlir/framework/op.h"
 #include "paddle/cinn/hlir/framework/op_strategy.h"
 #include "paddle/cinn/hlir/op/contrib/reciprocal.h"
