+++ paddle/cinn/hlir/op/contrib/triangular_solve.cc
@@ -17,7 +17,6 @@
 
 #include "paddle/cinn/common/common.h"
 #include "paddle/cinn/common/macros.h"
-#include "paddle/cinn/hlir/framework/node.h"
 #include "paddle/cinn/hlir/framework/op.h"
 #include "paddle/cinn/hlir/framework/op_strategy.h"
 #include "paddle/cinn/hlir/op/op_util.h"
