+++ paddle/cinn/hlir/op/nn.cc
@@ -17,7 +17,6 @@
 #include <functional>
 
 #include "paddle/cinn/adt/op_equation_context.h"
-#include "paddle/cinn/hlir/framework/node.h"
 #include "paddle/cinn/hlir/framework/op.h"
 #include "paddle/cinn/hlir/framework/op_strategy.h"
 #include "paddle/cinn/hlir/op/op_util.h"
