+++ paddle/cinn/hlir/op/reduction.cc
@@ -19,7 +19,6 @@
 #include <vector>
 
 #include "paddle/cinn/adt/op_equation_context.h"
-#include "paddle/cinn/hlir/framework/node.h"
 #include "paddle/cinn/hlir/framework/op.h"
 #include "paddle/cinn/hlir/framework/op_strategy.h"
 #include "paddle/cinn/hlir/pe/broadcast.h"
