+++ paddle/cinn/hlir/op/transform.cc
@@ -17,7 +17,6 @@
 #include <algorithm>
 
 #include "paddle/cinn/common/cas.h"
-#include "paddle/cinn/hlir/framework/node.h"
 #include "paddle/cinn/hlir/framework/op.h"
 #include "paddle/cinn/hlir/framework/op_strategy.h"
 #include "paddle/cinn/hlir/op/op_util.h"
