+++ paddle/cinn/hlir/pe/ir_schedule_pe.h
@@ -19,7 +19,6 @@
 #include <string>
 #include <vector>
 
-#include "paddle/cinn/hlir/framework/node.h"
 #include "paddle/cinn/hlir/pe/schedule_param.pb.h"
 #include "paddle/cinn/ir/ir.h"
 #include "paddle/cinn/ir/schedule/ir_schedule.h"
